backend/requirements.txt
pandas
Flask
Flask-Cors
openvino-dev
numpy
opencv-python
pytest
